requirements.txt
streamlit
Pillow
pandas
snowflake-connector-python